requirements.txt
torch>=2.1
torchvision>=0.16
numpy
pyyaml
tqdm
timm
transformers
datasets
